step9

+++ controllers/blogs.js
@@ -92,8 +92,26 @@ blogsRouter.post('/', async (request, response/*, next*/) => {
 
 blogsRouter.delete('/:id', async (request, response/*, next*/) => {
 
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user && blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'Unauthorized user' })
+  }
+
+
+
+
+
 
   /*
     Blog.findByIdAndRemove(request.params.id)
